requirements.txt
fastapi
uvicorn
streamlit
openai
pillow
numpy
matplotlib
rasterio
python-dotenv
requests
python-multipart